// websites/abaatechsolutions/src/includes/types.ts
export interface DBSettingsSchema {
    user?: string;
    password?: string;
    host?: string;
    port?: number;
    database?: string;
}

export interface DBResponseSchema {
    status: boolean;
    message: string;
    data: [key: any][] | [key: any]
}

export interface UserRoleSchema {
    name: UserRolesType;
    description: string;
    permissions: PrivilegesType[];
}

export type  UserRolesType = "Guest" | "Subscriber" | "Author" | "Editor" | "Administrator";
export type PrivilegesType = "edit" | "delete" | "view" | "comment";
// ...
export interface EmailsSchema {
    default: {
        email?: string;
        password?: string;
    },
    support: {
        email?: string;
        password?: string;
    },
    noreply: {
        email?: string;
        password?: string;
    }
}

export interface MailServerSchema {
    host: string | undefined;
    port: number | undefined;
    secure: boolean | undefined;
    authUser: string | undefined;
    authPass: string | undefined;
}